requirements.txt
kafka-python==2.0.2
pyflink==1.18.0
requests==2.31.0
